requirements.txt
Django>=5.1
djangorestframework>=3.15
django-filter>=24.0
whitenoise>=6.5
reportlab>=4.0
